+++ modules/Updater/bootstrap.php
@@ -3,6 +3,14 @@
 // Register Helpers
 $this->helpers['updater'] = 'Updater\\Helper\\Updater';
 
+
+// load admin related code
+$this->on('app.admin.init', function() {
+
+    include(__DIR__.'/admin.php');
+
+}, -1000);
+
 // load cli related code
 $this->on('app.cli.init', function($cli) {
     $app = $this;